1. add re-generate code if error occur, max 3 time; 2. add error handle for AI Chat;

+++ llm/llm_setup.py
@@ -12,20 +12,28 @@ class OpenAIAzure():
 
         self.WORK_ENV_DIR = './'
         self.ENV_FILE = 'key.txt'
+        self.config_details = {}
     def setup_env(self):
-        shutil.copyfile(os.path.join(self.WORK_ENV_DIR, self.ENV_FILE), ".env")
-        load_dotenv()
+        # Load OpenAI key
+        if os.path.exists(os.path.join(self.WORK_ENV_DIR, self.ENV_FILE)):
+            shutil.copyfile(os.path.join(self.WORK_ENV_DIR, self.ENV_FILE), ".env")
+            load_dotenv()
+        else:
+            raise APIKeyNotFoundError("key.txt with OpenAI API is required")
 
         # Load config values
-        with open(r'config.json') as config_file:
-            self.config_details = json.load(config_file)
+        if os.path.exists(os.path.join(r'config.json')):
+            with open(r'config.json') as config_file:
+                self.config_details = json.load(config_file)
 
-        # Setting up the embedding model
-        embedding_model_name = self.config_details['EMBEDDING_MODEL']
-        openai.api_type = "azure"
-        openai.api_base = self.config_details['OPENAI_API_BASE']
-        openai.api_version = self.config_details['EMBEDDING_MODEL_VERSION']
-        openai.api_key = os.getenv("OPENAI_API_KEY")
+            # Setting up the embedding model
+            embedding_model_name = self.config_details['EMBEDDING_MODEL']
+            openai.api_type = "azure"
+            openai.api_base = self.config_details['OPENAI_API_BASE']
+            openai.api_version = self.config_details['EMBEDDING_MODEL_VERSION']
+            openai.api_key = os.getenv("OPENAI_API_KEY")
+        else:
+            raise AzureConfigNotFoundError("config.json with Azure OpenAI config is required")
 
     def create_chat_model(self):
         # max LLM token input size
@@ -42,3 +50,20 @@ class OpenAIAzure():
                                max_tokens=num_output,
                                temperature=0.5,
                                )
+
+
+class APIKeyNotFoundError(Exception):
+    """
+    Raised when the API key is not defined/declared.
+
+    Args:
+        Exception (Exception): APIKeyNotFoundError
+    """
+
+class AzureConfigNotFoundError(Exception):
+    """
+    Raised when the API key is not defined/declared.
+
+    Args:
+        Exception (Exception): APIKeyNotFoundError
+    """
